Added course selection for the grades template
Had to move courses from a regular js import to local storage. This was because of issues importing it while also storing its filtered version for the catalog using redux.
This led to a large error where the localstorage is not loaded on the first load of the first page. Must be resolved in the future.